--- src/main/java/com/sprint/mission/discodeit/security/handler/Http403ForbiddenAccessDeniedHandler.java
@@ -1,32 +1,0 @@
-package com.sprint.mission.discodeit.security.handler;
-
-import java.io.IOException;
-
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sprint.mission.discodeit.exception.ErrorResponse;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public class Http403ForbiddenAccessDeniedHandler implements AccessDeniedHandler {
-
-	private final ObjectMapper objectMapper;
-
-	@Override
-	public void handle(HttpServletRequest request, HttpServletResponse response,
-		AccessDeniedException accessDeniedException) throws IOException, ServletException {
-		response.setStatus(HttpServletResponse.SC_FORBIDDEN);
-		response.setContentType("application/json");
-		response.setCharacterEncoding("UTF-8");
-
-		ErrorResponse errorResponse = new ErrorResponse(accessDeniedException,
-			HttpServletResponse.SC_FORBIDDEN);
-		response.getWriter().write(objectMapper.writeValueAsString(errorResponse));
-	}
-}
