--- src/main/java/com/sprint/mission/discodeit/security/handler/HttpStatusReturningLogoutSuccessHandler.java
@@ -1,20 +1,0 @@
-package com.sprint.mission.discodeit.security.handler;
-
-import java.io.IOException;
-
-import org.springframework.security.core.Authentication;
-import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class HttpStatusReturningLogoutSuccessHandler implements LogoutSuccessHandler {
-
-	@Override
-	public void onLogoutSuccess(HttpServletRequest request, HttpServletResponse response,
-		Authentication authentication) throws IOException, ServletException {
-
-		response.setStatus(HttpServletResponse.SC_NO_CONTENT);
-	}
-}
