+++ src/main/java/com/sprint/mission/discodeit/security/jwt/JwtTokenProvider.java
@@ -18,7 +18,7 @@ import com.nimbusds.jose.crypto.MACVerifier;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
 import com.sprint.mission.discodeit.dto.data.UserDto;
-import com.sprint.mission.discodeit.security.dto.DiscodeitUserDetails;
+import com.sprint.mission.discodeit.security.userdetails.DiscodeitUserDetails;
 
 import jakarta.servlet.http.Cookie;
