+++ src/main/java/com/sprint/mission/discodeit/security/userdetails/DiscodeitUserDetailsService.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.security.service;
+package com.sprint.mission.discodeit.security.userdetails;
 
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -10,7 +10,6 @@ import com.sprint.mission.discodeit.dto.data.UserDto;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.mapper.UserMapper;
 import com.sprint.mission.discodeit.repository.UserRepository;
-import com.sprint.mission.discodeit.security.dto.DiscodeitUserDetails;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
